Unify Measure newtype interiors to Int, remove ~70 ceremony fromIntegral calls
On 64-bit GHC (base >= 4.18), Int and Int64 have identical range.
Offset, FileSize, and Delta wrapped Int64 while Length and Position
wrapped Int, forcing fromIntegral at every interaction point. Change
all three to Int, removing the ceremony and keeping fromIntegral only
at real serialization boundaries (Int64 wire formats, Word32/Word16
encoding, Integer for hSeek/hSetFileSize).

+++ app/Main.hs
@@ -759,9 +759,9 @@ warnFileSize (FileSize expectedSize) (FileSize actualSize) =
 
 warnSourceBytes :: String -> Offset -> ByteString.ByteString -> ByteString.ByteString -> IO ()
 warnSourceBytes label checkOffset expectedData sourceBytes =
-  let actual = safeSlice (fromIntegral (unOffset checkOffset)) (ByteString.length expectedData) sourceBytes
+  let actual = safeSlice (unOffset checkOffset) (ByteString.length expectedData) sourceBytes
   in when (actual /= expectedData) $
-       warn (label ++ " mismatch at 0x" ++ padHex 8 (fromIntegral (unOffset checkOffset)))
+       warn (label ++ " mismatch at 0x" ++ padHex 8 (unOffset checkOffset))
 
 safeSlice :: Int -> Int -> ByteString.ByteString -> ByteString.ByteString
 safeSlice offset sliceLength input = ByteString.take sliceLength (ByteString.drop offset input)
